python-transcript-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==1.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
cryptography==41.0.7
yt-dlp>=2024.12.13
deepgram-sdk==4.5.0
requests==2.31.0
flask-swagger-ui==4.11.1
google-genai==0.8.0 